requirements.txt
astropy==7.2.0
astropy-iers-data==0.2026.2.16.0.48.25
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
filelock==3.24.3
fonttools==4.61.1
fsspec==2026.2.0
iniconfig==2.3.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
pyerfa==2.0.1.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
reportlab==4.4.10
scipy==1.17.0
setuptools==82.0.0
six==1.17.0
sympy==1.14.0
torch==2.10.0
typing_extensions==4.15.0


torch
torchvision
ultralytics
opencv-python
numpy
pandas
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-jaeger
opentelemetry-instrumentation
zlib